requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.31.0
beautifulsoup4>=4.13.0

# Information Retrieval
pyserini>=0.20.0
rank_bm25>=0.2.2
ir_datasets>=0.5.0
pytrec_eval>=0.5

# Text Processing
spacy>=3.3.0
nltk>=3.8.1
transformers>=4.11.0

# Neural Models
sentence-transformers>=2.2.0
torch>=1.10.0
faiss-cpu>=1.7.0

# Web Scraping
scrapy>=2.5.0
urllib3>=1.26.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.0.0

# Utilities
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
